dh_key_exchange/requirements.txt
pycryptodome==3.10.1
azure.identity==1.6.1
boto3==1.18.33
keepercommander==4.60
pytest==6.2.5
unittest2==1.1.0